+++ src/lib.rs
@@ -192,8 +192,32 @@ impl<T: Primitive> AtomicMaybeUninit<T> {
         }
     }
 
-    // TODO: Add from_ptr once it is stable on std atomic types.
-    // https://github.com/rust-lang/rust/issues/108652
+    /// Creates a new reference to an atomic value from a pointer.
+    ///
+    /// # Safety
+    ///
+    /// * `ptr` must be aligned to `align_of::<AtomicMaybeUninit<T>>()` (note that on some platforms this
+    ///   can be bigger than `align_of::<MaybeUninit<T>>()`).
+    /// * `ptr` must be [valid] for both reads and writes for the whole lifetime `'a`.
+    /// * Non-atomic accesses to the value behind `ptr` must have a happens-before
+    ///   relationship with atomic accesses via the returned value (or vice-versa).
+    ///   * In other words, time periods where the value is accessed atomically may not
+    ///     overlap with periods where the value is accessed non-atomically.
+    ///   * This requirement is trivially satisfied if `ptr` is never used non-atomically
+    ///     for the duration of lifetime `'a`. Most use cases should be able to follow
+    ///     this guideline.
+    ///   * This requirement is also trivially satisfied if all accesses (atomic or not) are
+    ///     done from the same thread.
+    /// * This method should not be used to create overlapping or mixed-size atomic
+    ///   accesses, as these are not supported by the memory model.
+    ///
+    /// [valid]: core::ptr#safety
+    #[inline]
+    #[must_use]
+    pub unsafe fn from_ptr<'a>(ptr: *mut MaybeUninit<T>) -> &'a Self {
+        // SAFETY: guaranteed by the caller
+        unsafe { &*ptr.cast::<Self>() }
+    }
 
     /// Returns a mutable reference to the underlying value.
     ///